Refactor ARKG scalar helpers
