+++ Yubico.Core/src/Yubico/Core/Cryptography/ArkgByteUtilities.cs
@@ -1,0 +1,47 @@
+// Copyright 2025 Yubico AB
+//
+// Licensed under the Apache License, Version 2.0 (the "License").
+// You may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+using System;
+using System.Security.Cryptography;
+
+namespace Yubico.Core.Cryptography
+{
+    internal static class ArkgByteUtilities
+    {
+        internal static byte[] Sha256(byte[] input)
+        {
+            using var sha = SHA256.Create();
+            return sha.ComputeHash(input);
+        }
+
+        internal static byte[] Concat(params byte[][] parts)
+        {
+            int total = 0;
+            for (int i = 0; i < parts.Length; i++)
+            {
+                total += parts[i].Length;
+            }
+
+            byte[] result = new byte[total];
+            int offset = 0;
+            for (int i = 0; i < parts.Length; i++)
+            {
+                Buffer.BlockCopy(parts[i], 0, result, offset, parts[i].Length);
+                offset += parts[i].Length;
+            }
+
+            return result;
+        }
+    }
+}
