+++ Yubico.Core/src/Yubico/Core/Cryptography/ArkgPrimitivesOpenSsl.cs
@@ -13,13 +13,13 @@
 // limitations under the License.
 
 using System;
-using System.Globalization;
 using System.Numerics;
 using System.Security;
 using System.Security.Cryptography;
 using System.Text;
 using CommunityToolkit.Diagnostics;
 using Yubico.PlatformInterop;
+using static Yubico.Core.Cryptography.ArkgByteUtilities;
 
 namespace Yubico.Core.Cryptography
 {
@@ -38,16 +38,6 @@ namespace Yubico.Core.Cryptography
         private const int P256CoordinateLength = 32;
         private const int Sec1UncompressedLength = 1 + (2 * P256CoordinateLength);
         private const byte Sec1UncompressedTag = 0x04;
-
-        // DST_ext for ARKG-P256, draft-bradleylundberg-cfrg-arkg-10 section 4.1:
-        // https://www.ietf.org/archive/id/draft-bradleylundberg-cfrg-arkg-10.html#name-arkg-p256
-        private const string DstExt = "ARKG-P256";
-
-        // P-256 group order N (SEC 2 v2, section 2.4.2). Used for scalar reduction mod N.
-        private static readonly BigInteger N = BigInteger.Parse(
-            "00FFFFFFFF00000000FFFFFFFFFFFFFFFFBCE6FAADA7179E84F3B9CAC2FC632551",
-            NumberStyles.HexNumber,
-            CultureInfo.InvariantCulture);
 
         private enum PointValidationResult
         {
@@ -175,14 +165,14 @@ namespace Yubico.Core.Cryptography
         {
             byte[] dst = Concat(
                 Encoding.ASCII.GetBytes("ARKG-BL-EC."),
-                Encoding.ASCII.GetBytes(DstExt),
+                Encoding.ASCII.GetBytes(ArkgP256Scalar.DstExt),
                 context.ToArray());
-            return HashToScalar(inputKeyingMaterialTau, dst);
+            return ArkgP256Scalar.HashToScalar(inputKeyingMaterialTau, dst);
         }
 
         private static byte[] BlBlindPublicKey(ReadOnlySpan<byte> blindingPublicKey, BigInteger tau)
         {
-            byte[] tauBytes = ScalarToP256FixedWidthBytes(tau);
+            byte[] tauBytes = ArkgP256Scalar.ToFixedWidthBytes(tau);
             using SafeEcGroup group = NativeMethods.EcGroupNewByCurveName(
                 ECCurve.NamedCurves.nistP256.ToSslCurveId());
             using SafeEcPoint blindingPublicKeyPoint = Sec1ToPoint(group, blindingPublicKey);
@@ -218,7 +208,7 @@ namespace Yubico.Core.Cryptography
 
             (byte[] ephemeralPublicKey, BigInteger ephemeralSecretScalar) = KemDeriveKeypair(inputKeyingMaterial);
 
-            byte[] ephemeralSecretScalarBytes = ScalarToP256FixedWidthBytes(ephemeralSecretScalar);
+            byte[] ephemeralSecretScalarBytes = ArkgP256Scalar.ToFixedWidthBytes(ephemeralSecretScalar);
             byte[] kPrime = ComputeEcdhSharedSecret(ephemeralSecretScalarBytes, kemPublicKey);
 
             byte[] macInfo = Concat(
@@ -260,15 +250,15 @@ namespace Yubico.Core.Cryptography
         {
             byte[] dst = Concat(
                 Encoding.ASCII.GetBytes("ARKG-KEM-ECDH-KG.ARKG-ECDH."),
-                Encoding.ASCII.GetBytes(DstExt));
-            BigInteger secretScalar = HashToScalar(inputKeyingMaterial, dst);
+                Encoding.ASCII.GetBytes(ArkgP256Scalar.DstExt));
+            BigInteger secretScalar = ArkgP256Scalar.HashToScalar(inputKeyingMaterial, dst);
             byte[] publicKey = ScalarMulGenerator(secretScalar);
             return (publicKey, secretScalar);
         }
 
         private static byte[] ScalarMulGenerator(BigInteger scalar)
         {
-            byte[] scalarBytes = ScalarToP256FixedWidthBytes(scalar);
+            byte[] scalarBytes = ArkgP256Scalar.ToFixedWidthBytes(scalar);
             using SafeEcGroup group = NativeMethods.EcGroupNewByCurveName(
                 ECCurve.NamedCurves.nistP256.ToSslCurveId());
             using SafeEcPoint publicKeyPoint = NativeMethods.EcPointNew(group);
@@ -286,121 +276,6 @@ namespace Yubico.Core.Cryptography
             }
 
             return PointToSec1(group, publicKeyPoint);
-        }
-
-        // ---------------------------------------------------------------------
-        // Hash-to-scalar helpers
-        // ---------------------------------------------------------------------
-
-        private static BigInteger HashToScalar(ReadOnlySpan<byte> msg, ReadOnlySpan<byte> dst)
-        {
-            const int L = 48;
-            byte[] uniform = ExpandMessageXmdSha256(msg, dst, L);
-
-            // Follows RFC 9380 section 5.2 hash_to_field reduction with the
-            // P256_XMD:SHA-256_SSWU_RO_ section 8.2 parameters (m=1, L=48,
-            // expand_message_xmd/SHA-256), except ARKG-P256 hashes to a
-            // scalar by reducing modulo the P-256 group order N rather than
-            // the coordinate-field characteristic p.
-            return Os2Ip(uniform) % N;
-        }
-
-        // RFC 9380 expand_message_xmd instantiated with SHA-256 (b_in_bytes=32,
-        // s_in_bytes=64), as used by the P256_XMD:SHA-256 suites.
-        private static byte[] ExpandMessageXmdSha256(ReadOnlySpan<byte> msg, ReadOnlySpan<byte> dst, int lenInBytes)
-        {
-            const int BInBytes = 32;
-            const int SInBytes = 64;
-
-            // Equivalent to ceil(len_in_bytes / b_in_bytes) from RFC 9380.
-            int ell = (lenInBytes + BInBytes - 1) / BInBytes;
-            if (ell > 255 || lenInBytes > 65535 || dst.Length > 255)
-            {
-                throw new ArgumentException("expand_message_xmd parameter out of range.");
-            }
-
-            byte[] dstPrime = new byte[dst.Length + 1];
-            dst.CopyTo(dstPrime);
-            dstPrime[dst.Length] = (byte)dst.Length;
-
-            byte[] zPad = new byte[SInBytes];
-            // RFC 9380 uses I2OSP(len_in_bytes, 2); the range check above
-            // keeps the value representable in exactly two octets.
-            byte[] lIBStr = [(byte)((lenInBytes >> 8) & 0xFF), (byte)(lenInBytes & 0xFF)];
-
-            byte[] msgPrime = Concat(zPad, msg.ToArray(), lIBStr, [0x00], dstPrime);
-
-            byte[][] bVals = new byte[ell + 1][];
-            bVals[0] = Sha256(msgPrime);
-            bVals[1] = Sha256(Concat(bVals[0], [0x01], dstPrime));
-
-            Span<byte> xored = stackalloc byte[BInBytes];
-            for (int i = 2; i <= ell; i++)
-            {
-                for (int j = 0; j < BInBytes; j++)
-                {
-                    xored[j] = (byte)(bVals[0][j] ^ bVals[i - 1][j]);
-                }
-
-                byte[] input = Concat(xored.ToArray(), [(byte)i], dstPrime);
-                bVals[i] = Sha256(input);
-            }
-
-            byte[] result = new byte[lenInBytes];
-            Span<byte> resultSpan = result;
-            int offset = 0;
-            for (int i = 1; i <= ell && offset < lenInBytes; i++)
-            {
-                int copy = Math.Min(BInBytes, lenInBytes - offset);
-                bVals[i].AsSpan(0, copy).CopyTo(resultSpan[offset..]);
-                offset += copy;
-            }
-
-            return result;
-        }
-
-        // ---------------------------------------------------------------------
-        // Conversion helpers
-        // ---------------------------------------------------------------------
-
-        private static byte[] ScalarToP256FixedWidthBytes(BigInteger scalar)
-        {
-            // Local fixed-width P-256 scalar encoding: BigInteger is signed
-            // little-endian, while OpenSSL BN input expects unsigned big-endian.
-            byte[] le = scalar.ToByteArray();
-            int len = le.Length;
-            if (len > 1 && le[len - 1] == 0)
-            {
-                len--;
-            }
-
-            byte[] be = new byte[P256CoordinateLength];
-            int copy = Math.Min(len, P256CoordinateLength);
-            for (int i = 0; i < copy; i++)
-            {
-                be[P256CoordinateLength - 1 - i] = le[i];
-            }
-
-            return be;
-        }
-
-        // RFC 8017 OS2IP: interpret an octet string as a non-negative integer.
-        private static BigInteger Os2Ip(ReadOnlySpan<byte> bytes)
-        {
-            byte[] padded = new byte[bytes.Length + 1];
-            for (int i = 0; i < bytes.Length; i++)
-            {
-                padded[bytes.Length - 1 - i] = bytes[i];
-            }
-
-            // padded[bytes.Length] = 0 by default — explicit positive sign.
-            return new BigInteger(padded);
-        }
-
-        private static byte[] Sha256(byte[] input)
-        {
-            using var sha = SHA256.Create();
-            return sha.ComputeHash(input);
         }
 
         private static SafeEcPoint Sec1ToPoint(SafeEcGroup group, ReadOnlySpan<byte> sec1)
@@ -448,25 +323,6 @@ namespace Yubico.Core.Cryptography
             xBytes.CopyTo(sec1Span[1..]);
             yBytes.CopyTo(sec1Span[(1 + P256CoordinateLength)..]);
             return sec1;
-        }
-
-        private static byte[] Concat(params byte[][] parts)
-        {
-            int total = 0;
-            for (int i = 0; i < parts.Length; i++)
-            {
-                total += parts[i].Length;
-            }
-
-            byte[] result = new byte[total];
-            int offset = 0;
-            for (int i = 0; i < parts.Length; i++)
-            {
-                Buffer.BlockCopy(parts[i], 0, result, offset, parts[i].Length);
-                offset += parts[i].Length;
-            }
-
-            return result;
         }
     }
 }
